+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "L13 Diff" extension will be documented in this file.
 
 ### Added
 - Added support for new extension [L13 Projects](https://marketplace.visualstudio.com/items?itemName=L13RARY.l13-projects)
-- Click on swap button also chnages the list view if paths are equal.
+- Alt + Click on swap button changes the list view, too.
 
 ## [0.20.0] - 2019-08-25